Just the basics and modified the gradle file to accomidate for the auth tools.
