+++ backend/utility/src/main/java/com/cmu/cps498/utility/users/User.java
@@ -4,36 +4,42 @@
  *  The user table will store the user's email, password, role, and isActive status--which will be useful for teachers
  */
 
-package com.cmu.cps498.users;
+package com.cmu.cps498.utility.users;
 
-@Entity //Marks the class as an JPA entity for mapping to a table
-@Getter // not sure what the package was for this. 
-@Setter // not sure what the package was for this.
+import jakarta.persistence.*; // JPA annotations for entity mapping
+import jakarta.validation.constraints.Email; // Email validation annotation
+import lombok.Getter; // Lombok annotation to generate getter methods
+import lombok.Setter; // Lombok annotation to generate setter methods
+import java.time.LocalDateTime; // Java 8 time API for timestamps
+
+@Entity // Marks the class as an JPA entity for mapping to a table
+@Getter
+@Setter
 public class User {
-    @Id //pirmary key
+    @Email // pirmary key
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    
+
     @Column(unique = true, nullable = false)
     private String email;
 
     @Column(nullable = false)
     private String password;
 
-    @Enumertaed(EnumType.STRING)
+    @Enumerated(EnumType.STRING)
     private Role role;
 
     @Column(nullable = false)
     private boolean isActive = true;
 
-
-    // might need these not sure yet. 
+    // might need these not sure yet.
     private LocalDateTime createdAt;
 
     private LocalDateTime updatedAt;
 
     private LocalDateTime lastLogin;
-    // logic for time stamps. 
+
+    // logic for time stamps.
     @PrePersist
     protected void onCreate() {
         this.createdAt = LocalDateTime.now();
@@ -41,9 +47,9 @@ public class User {
     }
 
     @PreUpdate
-    Protected void onUpdate() {
-        this.updateedAt = LocalDateTime.now();
+    protected void onUpdate() {
+        this.updatedAt = LocalDateTime.now();
     }
-    //-------- ^ Delete If not needed------------//
+    // -------- ^ Delete If not needed------------//
 
 }
